Fix two variants of LWJGL 3.3.1 with same name causing side effects
If you run MC 1.19 first, it will download the original LWJGL 3.3.1
which has not been patched for MC 1.16.5 and before. If you then run
MC 1.16.5, since the lib has the same name, the existing variant will
be reused, resulting in a crash. If you run MC 1.16.5 first and then
MC 1.19, the patched lib will be used for MC 1.19 instead, which is fine.

For simplicity, this commit just forces all versions to use the same
patched lib.